MP-1: fix
1) Merge base entity and named entity
2) Make START_SEQ private and add START_SEQ in tests
3) Add tests delete() in UserServiceTest
