+++ src/main/java/ru/marketplace/model/AbstractBaseEntity.java
@@ -10,19 +10,22 @@ import java.util.Objects;
 @MappedSuperclass
 @Access(AccessType.FIELD)
 public abstract class AbstractBaseEntity implements HasId {
-    public static final int START_SEQ = 100001;
-
+    private static final int START_SEQ = 100001;
 
     @Id
     @SequenceGenerator(name = "global_seq", sequenceName = "global_seq", allocationSize = 1, initialValue = START_SEQ)
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "global_seq")
     protected Integer id;
 
+    @Column(name = "name", nullable = false)
+    protected String name;
+
     protected AbstractBaseEntity() {
     }
 
-    protected AbstractBaseEntity(Integer id) {
+    protected AbstractBaseEntity(Integer id, String name) {
         this.id = id;
+        this.name = name;
     }
 
     public Integer getId() {
@@ -33,10 +36,19 @@ public abstract class AbstractBaseEntity implements HasId {
         this.id = id;
     }
 
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
     @Override
     public String toString() {
-        return getClass().getSimpleName() + ":" + id;
+        return getClass().getSimpleName() + ":" + id + '(' + name + ')';
     }
+
 
     @Override
     public boolean equals(Object o) {
@@ -45,11 +57,11 @@ public abstract class AbstractBaseEntity implements HasId {
             return false;
         }
         AbstractBaseEntity that = (AbstractBaseEntity) o;
-        return Objects.equals(id, that.id);
+        return id.equals(that.id) && name.equals(that.name);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(id);
+        return Objects.hash(id, name);
     }
 }
