--- src/main/java/ru/marketplace/model/AbstractNamedEntity.java
@@ -1,35 +1,0 @@
-package ru.marketplace.model;
-
-import javax.persistence.Column;
-import javax.persistence.Inheritance;
-import javax.persistence.InheritanceType;
-import javax.persistence.MappedSuperclass;
-
-@MappedSuperclass
-@Inheritance(strategy = InheritanceType.SINGLE_TABLE)
-public class AbstractNamedEntity extends AbstractBaseEntity{
-
-    @Column(name = "name", nullable = false)
-    protected String name;
-
-    protected AbstractNamedEntity() {
-    }
-
-    protected AbstractNamedEntity(Integer id, String name) {
-        super(id);
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public String toString() {
-        return super.toString() + '(' + name + ')';
-    }
-}
