+++ src/main/java/ru/marketplace/repository/user/CrudUserRepository.java
@@ -1,6 +1,7 @@
 package ru.marketplace.repository.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,6 +12,10 @@ import java.util.List;
 @Transactional(readOnly = true)
 public interface CrudUserRepository extends JpaRepository<User, Integer> {
 
+    @Transactional
+    @Modifying
+    @Query("DELETE FROM User u WHERE u.id=:id")
+    int delete(@Param("id") int id);
 
     @Query("SELECT u FROM User u JOIN FETCH u.products WHERE u.id=:userId")
     User getWithProducts(@Param("userId") int userId);
