+++ src/test/java/ru/marketplace/testdata/UserTestData.java
@@ -6,7 +6,8 @@ import ru.marketplace.model.User;
 
 import java.util.List;
 
-import static ru.marketplace.model.AbstractBaseEntity.START_SEQ;
+import static ru.marketplace.service.AbstractServiceTest.START_SEQ;
+
 
 public class UserTestData {
